Remove the dependency on JSR-375 apis from the api jar, and move the javax.security.enterprise.CallerPrincipal extension into the reference implementation.
